Added .env, README to twitter-scriper script; refactor for readability
