+++ scripts/twitter-scraper/.env
@@ -1,0 +1,2 @@
+CLIENT_ID = 'FILL ME IN'
+CLIENT_SECRET = 'FILL ME IN'
